冗長なキャスト削除

+++ include/others/util.hpp
@@ -76,7 +76,7 @@ class RandomGenerator final {
         const double min  = -std::numeric_limits<double>::infinity(),
         const double max  = std::numeric_limits<double>::infinity()
     ) noexcept -> double {
-        auto       dist = std::normal_distribution<double>{static_cast<double>(mean), div};
+        auto       dist = std::normal_distribution<double>{mean, div};
         const auto out  = dist(rng_);
         return std::clamp(out, min, max);
     }
